make-it-prototyping/rounds/round-02/classification/requirements.txt
torch>=1.9.0
torchvision>=0.10.0
efficientnet-pytorch>=0.7.1
pandas>=1.3.0
numpy>=1.21.0
matplotlib>=3.4.0
seaborn>=0.11.0
scikit-learn>=1.0.0
Pillow>=8.3.0
tqdm>=4.62.0
